sprint-5/step-2 added new structure for action-redicer services
