+++ src/pages/login/login.tsx
@@ -1,7 +1,7 @@
 import { EmailInput, PasswordInput, Button } from '@ya.praktikum/react-developer-burger-ui-components';
 import { Link } from 'react-router-dom';
 import { useSelector, useDispatch } from "react-redux";
-import { loginUser } from '../../services/actions/auth';
+import { loginUser } from '../../services/auth/action';
 import { useForm } from '../../hooks/useForm';
 import { ErrorBlock } from '../../local-uikit/components';
 import '../index.css';
