+++ src/services/reducer.ts
@@ -1,10 +1,10 @@
 import { combineReducers } from "@reduxjs/toolkit";
 
-import allIngredientsReducer from './all-ingredients-reducer';
-import burgerConstructorReducer from './burger-constructor-reducer';
-import orderReducer from './order-reducer';
+import allIngredientsReducer from './ingredients/reducer';
+import burgerConstructorReducer from './constructor/reducer';
+import orderReducer from './order/reducer';
 
-import {userReducer} from './auth';
+import {userReducer} from './auth/reducer';
 
 export const rootReducer = combineReducers({
     allIngredients: allIngredientsReducer,
